feat: `positivity` extension for `Real.log` of natural/integer casts and numeric literals (#5839)
This PR adds a positivity extension for expressions of the form `Real.log n` where `n` is a cast from a natural number or an integer. (Since `positivity` can't handle conditions like `1 ≤ x`, this is pretty much the best we can do for the log.)

Also, the namespace of the positivity extension for `exp` is corrected.