feat: make `lake exe cache get!` unpack everything (#5991)
Previously `lake exe cache get!` would download all files from the server, which fixes corrupted files in the `.cache` directory, but it would still skip files up to date according to the `.trace` files, so corrupted files in the `build` directory would not be fixed. (Before `leantar` all files would be unpacked regardless.) Now all files are unpacked when using `get!`. (This could be a separate option, but `get!!` is probably too confusing.)