doc: merge mathlib3 and mathlib4 `README`s (#6160)
This tries to merge the `README`s for mathlib3 and mathlib4 in a sensible way. Some irrelevant material about `leanproject` was removed, as well as a few mathlib3-specific items. Some of these should be replaced by mathlib4 counterparts in the future.

A section has been added about transitioning from Lean 3 to Lean 4 linking to the survival guide and running mathport on a project depending on mathlib.