refactor(Probability/Kernel/Composition): define compProd for all kernels (#6054)
`compProd` and `prod` make sense only for s-finite kernels and were defined only for those. I define them for all kernels, with default value 0 when one of the kernels is not s-finite. The new definitions allow rewriting a kernel inside a compProd, which was not possible before (it would raise a "motive is not type correct" error due to the s-finite argument).



Co-authored-by: Rémy Degenne <remydegenne@gmail.com>