feat: allow differential objects with shifts in an AddMonoidWithOne (#6246)
What I was really hoping to do was show that ℕ-graded DGOs were the same as ℕ-graded chain complexes (we have the ℤ version), but this still seems awkward. If anyone would like to help with that, please let me know. I think this might be helpful in constructor tensor products of chain complexes with less suffering.



Co-authored-by: Scott Morrison <scott.morrison@gmail.com>