feat(Data.Set.Ncard): enat set cardinality (#5908)
This PR is a second attempt at defining the cardinality of a set as extended natural number, with a function `encard`. The implementation involves a refactor, where the existing`ncard` function is redefined in terms of `encard`. This shortens a lot of proofs, reduces reliance on the `Finset` API, and allows for a potential future refactor where `ncard` is removed if it is decided to be redundant.  



Co-authored-by: Jireh Loreaux <loreaujy@gmail.com>