chore: remove some redundant `Nontrivial R` assumptions.
The real fix is to change the definition of `Module.Free.chooseBasisIndex` and
`Module.Free.chooseBasis`. The problem is that if `R` is trivial then there is
no guarantee that the chosen basis has finite indexing set. A possible fix would
be to change `Module.Free.chooseBasisIndex` to:
```lean
def ChooseBasisIndex :=
  if Nontrivial R then (exists_basis (R := R) (M := M)).some.1 else PUnit
```