refactor: use junk values in Submodule.toLinearPMap (#5529)
Change the definition of `Submodule.toLinearPMap` to use a junk value in the case that the condition that the subspace defines the graph of a function is not satisfied. In this case we define `Submodule.toLinearPMap` as the zero map. The domain is the same so that the domain does not depend on the graph-condition.



Co-authored-by: Oliver Nash <github@olivernash.org>