feat: the product of Borel spaces is Borel when either of them is second-countable (#6689)
We have currently that the product of two Borel spaces is Borel when both of them are second-countable. It is in fact sufficient to assume that only one of them is second-countable. We prove this in this PR.

Also move the definition of `SecondCountableEither` from `Function.StronglyMeasurable` to `BorelSpace.Basic` to be able to use it in the statement of the above theorem.