feat: Sup-closed sets (#6901)
This defines sets closed under supremum/infimum, shows that every set has a sup-closure/inf-closure and prove that if every sup-closed/inf-closed set in a sup-semilattice/inf-semilattice has a least upper bound/greatest lower, then the lattice is in fact complete.

As a bonus, we use our new predicate in `Order.CompactlyGenerated`.

Co-authored-by: Christopher Hoskin <christopher.hoskin@gmail.com>