feat: auto-suggest style fixes in CI (#6569)
This PR tries to help with the issue of contributors pushing a branch, closing their computer, and then later getting notified about a trivial style issue like, leaving a trailing space or a `set_option pp.proofs true` on somewhere.

With this change, the style linter output is added as github suggestions to your PR, which can be easily applied from the web interface without having to check out the branch again.

These checks run in a separate workflow to the original linters as the suggestions can only be added to PRs, not to branches, so this workflow will not run on branches not yet PRed. This is a slight duplication of work when a branch is part of a PR, but the time is so minimal compared to actually building / tactic linting that we dont think this will have a noticable effect on CI resources. 

The check all files imported suggesters are less likely to trigger if the user simply forgot the file. This is due to a technical limitation of github suggestions; suggestions can only be added to lines modified in the diff. It will at least fix help things if the user alphabetises wrongly (this has actually happened to me before...)
We _could_ have the linters automatically push fixes, but that is more invasive and potentially annoying so we hold off that for now until there is appetite for it.

We also moved the slightly expensive install bibtool step later as that step takes some time compared to the others and its nice to get more instant feedback from the earlier steps if they are going to fail anyway.



Co-authored-by: Julian Berman <Julian@GrayVines.com>