feat: better junk value for fderivWithin at isolated points (#7117)
Currently, when `x` is isolated in the set `s`, then `fderivWithin k f s x` can be anything. We modify the definition by ensuring that it is equal to `0` in this case. This ensures that the range of `derivWithin` is always contained in the closure of the span of the range of `f`.