perf: remove overspecified fields (#6965)
This removes redundant field values of the form `add := add` for smaller terms and less unfolding during unification. 

A list of all files containing a structure instance of the form `{ a1, ... with x1 := val, ... }` where some `xi` is a field of some `aj` was generated by modifying the structure instance elaboration algorithm to print such overlaps to stdout in a custom toolchain. 

Using that toolchain, I went through each file on the list and attempted to remove algebraic fields that overlapped and were redundant, eg `add := add` and not `toFun` (though some other ones did creep in). If things broke (which was the case in a couple of cases), I did not push further and reverted. 

It is possible that pushing harder and trying to remove all redundant overlaps will yield further improvements.