feat: pp.beta
I do not recommend using this feature, but since Lean 3 had it, and it's possible for us to implement it as mathlib code, here it is.

This is future-proof against core implementing it in the sense that if it gets implemented there will likely be a name collision, and then these two modules can be deleted.