chore(CategoryTheory): swapping the names of opOp and unopUnop (#7344)
This PR swaps the names of `opOp` and `unopUnop` which are the functors of the equivalence of categories `opOpEquivalence : Cᵒᵖᵒᵖ ≌ C`. As `Opposite.op` is the map `C → Cᵒᵖ`, the functor `C ⥤ Cᵒᵖᵒᵖ` should be named `opOp` and the functor `Cᵒᵖᵒᵖ ⥤ C` should be named `unopUnop`. It was previously the opposite; this PR fixes this.