fix: rm off-by-one error in finset.eventually_constant_prod (#7333)
The hypothesis is that the function is trivial for all indices `\geq N`, and the original conclusion is that for `n \geq N` the product up to `n` is equal to the product up to `N` (namely `range N+1`), but we can strengthen this by removing the index `N` part.  This change is especially important for simplifying products and sums with only the initial term nontrivial.