refactor: remove some Aesop calls in Category.Basic
This is in preparation for the next Aesop version, which will include a
change to the handling of simp theorems. That change can affect the
order in which simp lemmas are applied. Some Aesop calls in
CategoryTheory.Category.Basic were sensitive to this ordering, so I've
replaced them with simpler proofs.