refactor: purge aesop_cat_nonterminal (#7505)
`aesop_cat_nonterminal` is a non-terminal variant of `aesop`. It's not supposed to be used in production code since it's even worse than non-terminal `simp`. However, there were a few occurrences left (presumably from the port), which this PR removes.

The only nontrivial change is that I add mathlib's `rfl` tactic to the `CategoryTheory` Aesop rule set.