chore: remove many `Type _` before the colon (#7718)
We have turned to `Type*` instead of `Type _`, but many of them remained in mathlib because the straight replacement did not work. In general, having `Type _` before the colon is a code smell, though, as it hides which types should be in the same universe and which shouldn't, and is not very robust.

This PR replaces most of the remaining `Type _` before the colon (except those in category theory) by `Type*` or `Type u`. This has uncovered a few bugs (where declarations were not as polymorphic as they should be). 

I had to increase heartbeats at two places when replacing `Type _` by `Type*`, but I think it's worth it as it's really more robust.